feat: Add the ability to have different attributes for each map feature type
I've explicitly wanted to avoid using generics in the implementation. Using generics in this case would "litter" the code everywhere, whereas, now, we only have a tiny amount of casting done in consumer classes, and most of the heavily lifting is done in MapAttributesResolver.

It's also on purpose that the new attribute types do not have any custom values, rather, they only have defaults defined. I think it's sane to implement these custom attributes the first time we use them (so, in the territory area PR for area, and the first map path PR).